Replaced repitch with Xenakios repitchWDL and changed its formatting to match other functions

+++ libXCDP/include/Synthesis.h
@@ -9,7 +9,7 @@ namespace Synthesis {
 
 //wave is only evaluated from 0 to 2pi, which is its expected period
 
-Audio waveform( RealFunc wave, double length, RealFunc freq, int samplerate=44100, int oversample=16 );
+Audio waveform( RealFunc wave, double length, RealFunc freq, size_t samplerate = 44100, size_t oversample = 16 );
 
 Audio sine( double length, RealFunc freq );
 Audio square( double length, RealFunc freq );
